#332 add test for chainer frontend

+++ test/runtime/operators_test/lstm_test.py
@@ -13,8 +13,9 @@ from webdnn.graph.variables.constant_variable import ConstantVariable
 
 def _convert_to_chainer_order(x):
     # NOTE:
-    # In WebDNN, W_i, W_h, and b contains weights about input gate(v_i), forget gate(v_f) activated value(v_a), and output gate(v_o).
-    # However in chainer, they are packed as the order of v_a, v_i, v_f, and v_o
+    # In WebDNN, W_i, W_h, and b contains weights about input gate(v_i), forget gate(v_f) activated value(v_a), and output gate(v_o)
+    # based on this order (v_i, v_f, v_a, v_o). However in chainer, they are packed as different order (v_a, v_i, v_f, and v_o).
+    # Also, webdnn packs this weights as an tensor whose shape is (C1 or C2, 4, C2), but chainer packs as (C1 or C2, 4, C2)
     return x.reshape(x.shape[0], 4, x.shape[1] // 4).swapaxes(1, 2)[:, :, [2, 0, 1, 3]].reshape(x.shape)
 
 
@@ -23,12 +24,12 @@ def test_t_is_1():
     T = 1
     C1 = 128
     C2 = 64
-    vx = np.random.normal(size=(N, T, C1))
-    vw_input = np.random.normal(size=(C1, C2 * 4))
-    vw_hidden = np.random.normal(size=(C2, C2 * 4))
-    vb = np.random.normal(size=(C2 * 4,))
-    vc = np.zeros((N, C2))
-    vh = np.zeros((N, C2))
+    vx = np.random.normal(size=(N, T, C1)).astype(np.float32)
+    vw_input = np.random.normal(size=(C1, C2 * 4)).astype(np.float32)
+    vw_hidden = np.random.normal(size=(C2, C2 * 4)).astype(np.float32)
+    vb = np.random.normal(size=(C2 * 4,)).astype(np.float32)
+    vc = np.zeros((N, C2)).astype(np.float32)
+    vh = np.zeros((N, C2)).astype(np.float32)
 
     vw_input_c = _convert_to_chainer_order(vw_input)
     vw_hidden_c = _convert_to_chainer_order(vw_hidden)
@@ -59,12 +60,12 @@ def test_t_is_10():
     T = 10
     C1 = 128
     C2 = 64
-    vx = np.random.normal(size=(N, T, C1))
-    vw_input = np.random.normal(size=(C1, C2 * 4))
-    vw_hidden = np.random.normal(size=(C2, C2 * 4))
-    vb = np.random.normal(size=(C2 * 4,))
-    vc = np.zeros((N, C2))
-    vh = np.zeros((N, C2))
+    vx = np.random.normal(size=(N, T, C1)).astype(np.float32)
+    vw_input = np.random.normal(size=(C1, C2 * 4)).astype(np.float32)
+    vw_hidden = np.random.normal(size=(C2, C2 * 4)).astype(np.float32)
+    vb = np.random.normal(size=(C2 * 4,)).astype(np.float32)
+    vc = np.zeros((N, C2)).astype(np.float32)
+    vh = np.zeros((N, C2)).astype(np.float32)
 
     vw_input_c = _convert_to_chainer_order(vw_input)
     vw_hidden_c = _convert_to_chainer_order(vw_hidden)
